+++ cocos2dx/platform/mac/EAGLView.mm
@@ -327,7 +327,7 @@ static EAGLView *view;
 	float x = local_point.x;
 	float y = [self getHeight] - local_point.y;
 	
-    int ids[1] = {0};
+    intptr_t ids[1] = {0};
     float xs[1] = {0.0f};
     float ys[1] = {0.0f};
 	int tapcounts[1] = {0};
@@ -353,7 +353,7 @@ static EAGLView *view;
 	float x = local_point.x;
 	float y = [self getHeight] - local_point.y;
 
-    int ids[1] = {0};
+    intptr_t ids[1] = {0};
     float xs[1] = {0.0f};
     float ys[1] = {0.0f};
 	int tapcounts[1] = {0};
@@ -374,7 +374,7 @@ static EAGLView *view;
 	float x = local_point.x;
 	float y = [self getHeight] - local_point.y;
 
-    int ids[1] = {0};
+    intptr_t ids[1] = {0};
     float xs[1] = {0.0f};
     float ys[1] = {0.0f};
 	int tapcounts[1] = {0};
